+++ lib/components/OnlyClickListOptions.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import OnlyClickOption from './OnlyClickListOption';
 import classNames from 'classnames';
+import getDataTest from '../utils/dataTestHelper';
 
 function OnlyClickListOptions(props) {
   const {
@@ -19,6 +20,7 @@ function OnlyClickListOptions(props) {
     animatedSelection,
     withArrows,
     scrollable,
+    dataTest,
   } = props;
 
   return (
@@ -43,6 +45,7 @@ function OnlyClickListOptions(props) {
           optionComponent={optionComponent}
           animatedSelection={animatedSelection}
           withArrow={withArrows}
+          dataTest={dataTest || getDataTest(option.value)}
         />
       ))}
     </ul>
@@ -64,6 +67,7 @@ OnlyClickListOptions.propTypes = {
   optionComponent: PropTypes.func,
   withArrows: PropTypes.bool,
   scrollable: PropTypes.bool,
+  dataTest: PropTypes.string,
 };
 
 export default OnlyClickListOptions;