+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.37.0] - 2021-01-19
+
+## Added
+
+- New prop data-test in OnlyClickSelect, OnlyClickIconOptions, OnlyClickIconOption, IconRadioGroup, IconRadioInput, Toggle, AccordionSelect, ButtonsListSelect, ButtonsListSelectOption, ButtonsListSelectOptions, OnlyClickListOptions, OnlyClickMultiSelect, WideButton components to use it for Cypress tests
+
+## Changed
+
+- Updated node version from 8.10.0 to 10.15.0, updated circleci config to use new node version
+
 ## [2.36.6] - 2020-12-02
 
 ## Added
+
 - New prop to show labels instead of values for OnlyClickSelect Component